int id to long id

+++ src/main/java/Hyperlink.java
@@ -1,19 +1,19 @@
 import java.util.*
 
 public Hyperlink implements IHyperlink{
-    private int id;
+    private long id;
     private String name, url;
     private List<String> tags = new List<String>();
-    public Hyperlink (int id, String name, String url, List<String> tags) {
+    public Hyperlink (long id, String name, String url, List<String> tags) {
 	this.id = id;
 	this.name = name;
 	this.url = url;
 	this.tags = new ArrayList<String>(tags);
     }
-    public void setId(int id){
+    public void setId(long id){
 	this.id = id;
     }
-    public int getId() {
+    public long getId() {
 	return id;
     }
     public void setName(String name){
